requirements.txt
flask
flask-smorest
python-dotenv
google-genai
pydantic
flask_cors
web3